Move events to plugins and core
