+++ client/events.go
@@ -21,7 +21,7 @@ import (
 
 	eventsapi "github.com/containerd/containerd/v2/api/services/events/v1"
 	"github.com/containerd/containerd/v2/api/types"
-	"github.com/containerd/containerd/v2/pkg/events"
+	"github.com/containerd/containerd/v2/core/events"
 	"github.com/containerd/containerd/v2/protobuf"
 	"github.com/containerd/errdefs"
 	"github.com/containerd/typeurl/v2"
