+++ core/metrics/cgroups/cgroups.go
@@ -22,10 +22,10 @@ import (
 	"context"
 
 	"github.com/containerd/cgroups/v3"
+	"github.com/containerd/containerd/v2/core/events"
 	v1 "github.com/containerd/containerd/v2/core/metrics/cgroups/v1"
 	v2 "github.com/containerd/containerd/v2/core/metrics/cgroups/v2"
 	"github.com/containerd/containerd/v2/core/runtime"
-	"github.com/containerd/containerd/v2/pkg/events"
 	"github.com/containerd/containerd/v2/plugins"
 	"github.com/containerd/containerd/v2/version"
 	"github.com/containerd/platforms"
