+++ core/runtime/v2/shim.go
@@ -38,9 +38,9 @@ import (
 	eventstypes "github.com/containerd/containerd/v2/api/events"
 	task "github.com/containerd/containerd/v2/api/runtime/task/v3"
 	"github.com/containerd/containerd/v2/api/types"
+	"github.com/containerd/containerd/v2/core/events/exchange"
 	"github.com/containerd/containerd/v2/core/runtime"
 	client "github.com/containerd/containerd/v2/core/runtime/v2/shim"
-	"github.com/containerd/containerd/v2/pkg/events/exchange"
 	"github.com/containerd/containerd/v2/pkg/identifiers"
 	"github.com/containerd/containerd/v2/pkg/timeout"
 	"github.com/containerd/containerd/v2/protobuf"
