+++ core/runtime/v2/shim/publisher.go
@@ -23,7 +23,7 @@ import (
 
 	v1 "github.com/containerd/containerd/v2/api/services/ttrpc/events/v1"
 	"github.com/containerd/containerd/v2/api/types"
-	"github.com/containerd/containerd/v2/pkg/events"
+	"github.com/containerd/containerd/v2/core/events"
 	"github.com/containerd/containerd/v2/pkg/namespaces"
 	"github.com/containerd/containerd/v2/pkg/ttrpcutil"
 	"github.com/containerd/containerd/v2/protobuf"
