+++ plugins/services/containers/local.go
@@ -23,8 +23,8 @@ import (
 	eventstypes "github.com/containerd/containerd/v2/api/events"
 	api "github.com/containerd/containerd/v2/api/services/containers/v1"
 	"github.com/containerd/containerd/v2/core/containers"
+	"github.com/containerd/containerd/v2/core/events"
 	"github.com/containerd/containerd/v2/core/metadata"
-	"github.com/containerd/containerd/v2/pkg/events"
 	"github.com/containerd/containerd/v2/plugins"
 	"github.com/containerd/containerd/v2/plugins/services"
 	ptypes "github.com/containerd/containerd/v2/protobuf/types"
