+++ plugins/services/events/service.go
@@ -23,8 +23,8 @@ import (
 	api "github.com/containerd/containerd/v2/api/services/events/v1"
 	apittrpc "github.com/containerd/containerd/v2/api/services/ttrpc/events/v1"
 	"github.com/containerd/containerd/v2/api/types"
-	"github.com/containerd/containerd/v2/pkg/events"
-	"github.com/containerd/containerd/v2/pkg/events/exchange"
+	"github.com/containerd/containerd/v2/core/events"
+	"github.com/containerd/containerd/v2/core/events/exchange"
 	"github.com/containerd/containerd/v2/plugins"
 	"github.com/containerd/containerd/v2/protobuf"
 	ptypes "github.com/containerd/containerd/v2/protobuf/types"
