+++ plugins/services/sandbox/controller_service.go
@@ -26,8 +26,8 @@ import (
 
 	eventtypes "github.com/containerd/containerd/v2/api/events"
 	api "github.com/containerd/containerd/v2/api/services/sandbox/v1"
+	"github.com/containerd/containerd/v2/core/events"
 	"github.com/containerd/containerd/v2/core/sandbox"
-	"github.com/containerd/containerd/v2/pkg/events"
 	"github.com/containerd/containerd/v2/plugins"
 	"github.com/containerd/containerd/v2/protobuf"
 	"github.com/containerd/errdefs"
